+++ src/main/java/com/upublic/vo/Province.java
@@ -1,14 +1,14 @@
 package com.upublic.vo;
 
 public class Province {
-  private Long pid;
+  private Integer pid;
   private String pname;
 
-  public Long getPid() {
+  public Integer getPid() {
     return pid;
   }
 
-  public void setPid(Long pid) {
+  public void setPid(Integer pid) {
     this.pid = pid;
   }
